requirements.txt
fastapi>=0.109.1,<0.110.0
uvicorn>=0.27.1,<0.28.0
python-dotenv>=1.0.1,<2.0.0
httpx>=0.26.0,<0.27.0
pydantic>=2.6.1,<3.0.0 